build: update sbt and using sbt-tpolcat
* Update sbt to 1.4.6
* Update scala to 2.13.4
* Start using
[`sbt-tpolcat`](https://github.com/DavidGregory084/sbt-tpolecat)

+++ src/main/scala/com/raquo/airstream/eventstream/DelayEventStream.scala
@@ -18,7 +18,7 @@ class DelayEventStream[A](
 
   override protected[airstream] def onNext(nextValue: A, transaction: Transaction): Unit = {
     var timerHandle: SetTimeoutHandle = null
-    timerHandle = js.timers.setTimeout(delayMillis) {
+    timerHandle = js.timers.setTimeout(delayMillis.toDouble) {
       //println(s"> init trx from DelayEventStream.onNext($nextValue)")
       timerHandles.splice(timerHandles.indexOf(timerHandle), deleteCount = 1) // Remove handle
       new Transaction(fireValue(nextValue, _))
@@ -28,7 +28,7 @@ class DelayEventStream[A](
 
   override def onError(nextError: Throwable, transaction: Transaction): Unit = {
     var timerHandle: SetTimeoutHandle = null
-    timerHandle = js.timers.setTimeout(delayMillis) {
+    timerHandle = js.timers.setTimeout(delayMillis.toDouble) {
       timerHandles.splice(timerHandles.indexOf(timerHandle), deleteCount = 1) // Remove handle
       new Transaction(fireError(nextError, _))
     }
